+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.97.0
 
 * `[passthru]` Added package with pass-thru reader and writer
+* `[progress]` Migrate to `passthru` package
 * `[fmtutil/table]` Improved borders and separators rendering
 * `[usage]` Improved environment info output
 * `[spinner]` Improved message rendering